### 2025-12-09 0f92da7
update bib
