+++ paper.bib
@@ -1,122 +1,89 @@
-% our paper
-@inproceedings{stevan2024performance,
-  author    = {Stevan, Antoine and Lavaur, Thomas and Lacan, J{\'e}r{\^o}me and Detchart, Jonathan and P{\'e}rennou, Tanguy},
-  title     = {Performance Evaluation of Polynomial Commitments for Erasure Code Based Information Dispersal},
-  booktitle = {10th International Conference on Information Systems Security and Privacy},
-  year      = {2024}
+@article{ambrona2023aplonk,
+    title        = {aPlonK: Aggregated PlonK from Multi-polynomial Commitment Schemes},
+    abstract     = {PlonK is a prominent universal and updatable zk-SNARK for general circuit satis ability. We present aPlonK, a variant of PlonK that reduces the proof size and veri cation time when multiple statements are proven in a batch. Both the aggregated proof size and the veri cation complexity of aPlonK are logarithmic in the number of aggregated statements. Our main building block, inspired by the techniques developed in SnarkPack (Gailly, Maller, Nitulescu, FC 2022), is a multi-polynomial commitment scheme, a new primitive that generalizes polynomial commitment schemes. Our techniques also include a mechanism for involving committed data into PlonK statements very e ciently, which can be of independent interest.},
+    language     = {en},
+    author       = {Ambrona, Miguel and Beunardeau, Marc and Schmitt, Anne-Laure and Toledo, Raphaël R},
+    booktitle    = {International Workshop on Security},
+    pages        = {195--213},
+    year         = {2023},
+    organization = {Springer},
 }
 
-% building blocks
+
+@article{boneh2020efficient,
+    title   = {Efficient polynomial commitment schemes for multiple points and polynomials},
+    author  = {Boneh, Dan and Drake, Justin and Fisch, Ben and Gabizon, Ariel},
+    journal = {Cryptology ePrint Archive},
+    year    = {2020},
+}
+
+
 @software{arkworks,
   author = {arkworks contributors},
   title  = {\texttt{arkworks} zkSNARK ecosystem},
   url    = {https://arkworks.rs},
-  year   = {2022}
+  year   = {2022},
 }
 
-@inproceedings{merkle1987digital,
-  title        = {A digital signature based on a conventional encryption function},
-  author       = {Merkle, Ralph C},
-  booktitle    = {Conference on the theory and application of cryptographic techniques},
-  pages        = {369--378},
-  year         = {1987},
-  organization = {Springer}
+
+@article{ethereum2024danksharding,
+  title = {What is Proto-Danksharding?},
+  url = {https://ethereum.org/roadmap/danksharding/#what-is-protodanksharding},
+  year = {2024},
 }
+
 
 @inproceedings{fiat1986prove,
-  title        = {How to prove yourself: Practical solutions to identification and signature problems},
-  author       = {Fiat, Amos and Shamir, Adi},
-  booktitle    = {Conference on the theory and application of cryptographic techniques},
-  pages        = {186--194},
-  year         = {1986},
-  organization = {Springer},
-  url          = {https://link.springer.com/content/pdf/10.1007/3-540-47721-7.pdf?pdf=button}
+    title        = {How to prove yourself: Practical solutions to identification and signature problems},
+    author       = {Fiat, Amos and Shamir, Adi},
+    booktitle    = {Conference on the theory and application of cryptographic techniques},
+    pages        = {186--194},
+    year         = {1986},
[…]
