### 2025-12-09 f56dc75
fix references in paper.md
