### 2025-12-09 #9577 @a-stevan
## finding DOIs (WIP ♻️)
- [`gitlab.isae-supaero.fr:komodo@0f92da72`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/0f92da72)
- [`gitlab.isae-supaero.fr:komodo@f56dc757`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/f56dc757)

| LaTeX tag                      | DOI                                                                              | Title                                                                                               |
| ------------------------------ | :------------------------------------------------------------------------------: | --------------------------------------------------------------------------------------------------- |
| `@ambrona2023aplonk`           |                                                                                  | _aPlonK: Aggregated PlonK from Multi-polynomial Commitment Schemes_                                 |
| `@boneh2020efficient`          |                                                                                  | _Efficient polynomial commitment schemes for multiple points and polynomials_                       |
| `@arkworks`                    | $\emptyset$                                                                      | _\texttt{arkworks} zkSNARK ecosystem_                                                               |
| `@ethereum2024danksharding`    | $\emptyset$                                                                      | _What is Proto-Danksharding?_                                                                       |
| `@fiat1986prove`               | [**10.1007/3-540-47721-7_12**](https://doi.org/10.1007/3-540-47721-7_12)         | _How to prove yourself: Practical solutions to identification and signature problems_               |
| `@gabizon2019plonk`            |                                                                                  | _Plonk: Permutations over lagrange-bases for oecumenical noninteractive arguments of knowledge_     |
| `@kate2010constant`            | [**10.1007/978-3-642-17373-8_11**](https://doi.org/10.1007/978-3-642-17373-8_11) | _Constant-size commitments to polynomials and their applications_                                   |
| `@merkle1987digital`           | [**10.1007/3-540-48184-2_32**](https://doi.org/10.1007/3-540-48184-2_32)         | _A digital signature based on a conventional encryption function_                                   |
| `@nazirkhanova2022information` | [**10.1145/3558535.3559778**](https://doi.org/10.1145/3558535.3559778)           | _Information dispersal with provable retrievability for rollups_                                    |
| `@rust-rse`                    | $\emptyset$                                                                      | _Rust implementation of Reed-Solomon erasure coding_                                                |
| `@stevan2023assessing`         | [**10.1007/978-3-031-89518-0_13**](https://doi.org/10.1007/978-3-031-89518-0_13) | _Assessing the Efficiency of Polynomial Commitment Schemes in Erasure Code-Based Data Distribution_ |
| `@stevan2024performance`       | [**10.5220/0012377900003648**](https://doi.org/10.5220/0012377900003648)         | _Performance Evaluation of Polynomial Commitments for Erasure Code Based Information Dispersal_     |

> [!note]
> the following references where not used and have been removed
> - _On the size of pairing-based non-interactive arguments_ (`@groth2016size`)
> - _Tezos lib_aplonk_ (`@tezos2024aplonk`)
> - _KZG in Practice: Polynomial Commitment Schemes and Their Usage in Scaling Ethereum_ (`@scroll2024`)
> - _avail: The essential base layer for modern blockchains._ (`@avail2024`)
