+++ paper.bib
@@ -7,6 +7,7 @@
     pages        = {195--213},
     year         = {2023},
     organization = {Springer},
+    doi          = {10.1007/978-3-031-41326-1_11},
 }
 
 
