### 2025-12-09 f66b494
add JOSS intro (adapted) to the README
