### 2025-12-09 ddf420e
update bib
