### 2025-12-09 07b161c
remove some sections (2561w -> 1596w)
