+++ paper.md
@@ -30,8 +30,6 @@ date: 24 November 2025
 bibliography: paper.bib
 ---
 
-\usetikzlibrary{shapes,arrows,positioning,calc}
-
 # Summary
 
 We present **Komodo**, a library that allows to encode data with erasure-code
@@ -175,174 +173,6 @@ interactive discussion where the _verifier_ imposes challenges to the _prover_
 and the latter tries to convince the former. This is not very practical and the
 implementation uses a technic known as the _Fiat-Shamir transform_ from
 [@fiat1986prove].
-
-## General data flow in **Komodo**
-
-This section shows a high-level overview of the data flow in **Komodo**. Some
-data $D$ is first encoded on the sender side. Then _commitments_ and _proofs_
-are computed. Eventually, on the receiver side, blocks are verified and decoded.
-Here is how the sender side works:
-
-\tikzset{
-    block/.style = {draw, fill=white, rectangle, minimum height=3em, minimum width=3em},
-    tmp/.style  = {coordinate},
-    sum/.style= {draw, fill=white, circle, node distance=1cm},
-    input/.style = {coordinate},
-    output/.style= {coordinate},
-    pinstyle/.style = {pin edge={to-,thin,black}}
-}
-
-\begin{tikzpicture}[auto, node distance=2cm,>=latex']
-    \node [block, fill=green!50] (data) {$D$};
-    \node [block, below of=data, fill=red!50] (source) {$(s_i)$};
-    \node [block, right of=source, node distance=4cm] (encoded) {$(e_j)$};
-    \node [block, right of=encoded, node distance=3cm, fill=yellow!20] (commitment) {$c$};
-    \node [block, below of=commitment, node distance=1.3cm, fill=yellow!30] (proof) {$(\pi_j)$};
-    \node [block, right of=commitment, node distance=3cm, fill=blue!50] (blocks) {$(b_j)$};
-    \coordinate (below_proof) at ([yshift=-3mm] proof.south);
-    \draw [->] (data) -- node{split into elements of $\mathbb{F}$} (source);
-    \draw [->] (source) -- node{\texttt{encode(k, n)}} (encoded);
-    \draw [->] (encoded) -- node[name=a,anchor=south]{\texttt{commit}} (commitment);
-    \draw [->] (a) |- node[anchor=north]{\texttt{prove}} (proof);
-    \draw [->] (commitment) -- node[name=b,anchor=south]{\texttt{build}} (blocks);
-    \draw [->] (proof) -| (b);
-    \draw [->] (encoded.south) |- (below_proof) -| (b);
-\end{tikzpicture}
-
-where
-
-- $S = (s_i)_{1 \leq i \leq k} \in \mathcal{M}_{m \times k}(\mathbb{F})$ is the
-  matrix of source shards
-- $M \in \mathcal{M}_{k \times n}(\mathbb{F})$ is the encoding matrix
-- $E = (e_j)_{1 \leq j \leq n} = S \times M \in \mathcal{M}_{m \times n}(\mathbb{F})$
-  is the matrix of encoded shards
-- $c \in \mathbb{F}$ is the common _commitment_
-- $(\pi_j)_{1 \leq j \leq n} \in \mathbb{F}^{n}$ are the proofs for each _shard_
-- $(b_j)_{1 \leq j \leq n}$ are the final proven blocks, i.e. $(e_j, c, \pi_j)$
-
-Once blocks are proven on the sender side, they can be dispersed, e.g. for
-distributed storage. During this dispersion process, they can be corrupted,
-either during transmission or on the storage nodes. Eventually, on the receiver
-side, they should be verified and decoded properly. Here is how the receiver
-side works:
-
-\begin{tikzpicture}[auto, node distance=2cm,>=latex']
[…]
