### 2025-12-09 2a5f982
simplify (1596w -> 1533w)
