### 2025-12-09 74497b2
move figure comments together
