### 2025-12-09 5ae7bd3
remove "komodo" section (1533w -> 972w)
