### 2025-12-09 a944850
add back protocols and arkworks (972w -> 1029w)
