### 2025-12-10 9aed094
homogenize use of "Komodo" and citations (1029w -> 1021w)
