### 2025-12-10 5b6e184
remove duplicate formulations (1021w -> 999w)
