### 2025-12-10 #9577 @a-stevan
@danielskatz i think our paper has been fixed as requested 😄

- DOIs have been added (there are 5 software libraries / blog posts / preprints that do not have any DOI)
- an introduction has been added to the README of Komodo
- the paper has been reduced from 2561 to 999 words!

let us know what the next step is, looking forward to discussing Komodo again!

## review changelog
* [`gitlab.isae-supaero.fr:dragoon/komodo@0f92da72`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/0f92da72) _update bib_
* [`gitlab.isae-supaero.fr:dragoon/komodo@f56dc757`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/f56dc757) _fix references in paper.md_
* [`gitlab.isae-supaero.fr:dragoon/komodo@a3db0002`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/a3db0002) _update bib_
* [`gitlab.isae-supaero.fr:dragoon/komodo@f66b494f`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/f66b494f) _add JOSS intro (adapted) to the README_
* [`gitlab.isae-supaero.fr:dragoon/komodo@ddf420e5`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/ddf420e5) _update bib_
* [`gitlab.isae-supaero.fr:dragoon/komodo@07b161c9`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/07b161c9) _remove some sections (2561w -> 1596w)_
* [`gitlab.isae-supaero.fr:dragoon/komodo@2a5f9826`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/2a5f9826) _simplify (1596w -> 1533w)_
* [`gitlab.isae-supaero.fr:dragoon/komodo@74497b24`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/74497b24) _move figure comments together_
* [`gitlab.isae-supaero.fr:dragoon/komodo@5ae7bd35`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/5ae7bd35) _remove "komodo" section (1533w -> 972w)_
* [`gitlab.isae-supaero.fr:dragoon/komodo@a9448503`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/a9448503) _add back protocols and arkworks (972w -> 1029w)_
* [`gitlab.isae-supaero.fr:dragoon/komodo@9aed0943`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/9aed0943) _homogenize use of "Komodo" and citations (1029w -> 1021w)_
* [`gitlab.isae-supaero.fr:dragoon/komodo@5b6e1843`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/5b6e1843) _remove duplicate formulations (1021w -> 999w)_
