### 2025-12-10 #9577 @danielskatz
👋 @a-stevan - thanks for the changes, but I see one problem that remains, as well as a new one I've just noticed.

First, I don't think you've added anything to the README to introduce the software.

Second, the repo where the software is located doesn't satisfy JOSS's requirements - see "The software must be hosted at a location where users can browse the source code files, open issues, and propose code changes without manual approval of (or payment for) accounts" in https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements  When I try to edit a file, I see this:
