### 2025-12-10 #9577 @a-stevan
> Ok, thanks - I wasn't looking at the right branch, and I will switch the review repo to the GitHub mirror...
>
amazing, this bot is pretty helpful !

(i was about to say that the branch is also `paper-joss` on the GitHub mirror, but it looks like the `set <URL> as repository` command took that into account)
