### 2025-12-15 9257fb2
change Semi-AVID API (dragoon/komodo!235)
while working on the benchmarks, i realized the API of Semi-AVID was incorrect and did not reflect the actual algorithm...

the other protocols are fine.

## changelog
- renames the `proof` field of `komodo::semi_avid::Block` to `commitment`
- renames `komodo::semi_avid::prove` to `komodo::semi_avid::commit`
- fixes and adapts all places where `komodo::semi_avid::prove` was used and usually stored in a `proof` variable

> [!note]
> additionally, this MR
> - refactors some "feature" imports
> - "_debug asserts_" that $k$ is equal to the length of the _l
[…]
