+++ src/semi_avid.rs
@@ -62,7 +62,7 @@
 //! let bytes = include_bytes!("../assets/dragoon_133x133.png").to_vec();
 //! # }
 //! ```
-//! - then, $\text{Semi-AVID}$ requires a trusted setup to prove and verify. This example shows a trusted
+//! - then, $\text{Semi-AVID}$ requires a trusted setup to commit and verify. This example shows a trusted
 //! setup big enough to support data as big as $10 \times 1024$ elements of $\mathbb{F}_p$, to
 //! allow users to reuse it with multiple files of varying lengths.
 //! ```
@@ -76,12 +76,12 @@
 //! let powers = komodo::zk::setup::<F, G>(10 * 1_024, &mut rng).unwrap();
 //! # }
 //! ```
-//! - we can now build an encoding matrix, encode the data, prove the shards and build [`Block`]s
+//! - we can now build an encoding matrix, encode the data, commit it, proving the shards implicitely, and build [`Block`]s
 //! ```
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, prove, verify};
+//! # use komodo::semi_avid::{build, commit, verify};
 //! #
 //! # fn main() {
 //! # let mut rng = ark_std::test_rng();
@@ -93,8 +93,8 @@
 //! #
 //! let encoding_mat = &komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
 //! let shards = komodo::fec::encode(&bytes, encoding_mat).unwrap();
-//! let proof = prove::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
-//! let blocks = build::<F, G, DP<F>>(&shards, &proof);
+//! let commitment = commit::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
+//! let blocks = build::<F, G, DP<F>>(&shards, &commitment);
 //! # }
 //! ```
 //! - finally, each [`Block`] can be verified individually, using the same trusted setup
@@ -102,7 +102,7 @@
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, prove, verify};
+//! # use komodo::semi_avid::{build, commit, verify};
 //! #
 //! # fn main() {
 //! # let mut rng = ark_std::test_rng();
@@ -114,8 +114,8 @@
 //! #
 //! # let encoding_mat = &komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
 //! # let shards = komodo::fec::encode(&bytes, encoding_mat).unwrap();
-//! # let proof = prove::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
-//! # let blocks = build::<F, G, DP<F>>(&shards, &proof);
+//! # let commitment = commit::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
+//! # let blocks = build::<F, G, DP<F>>(&shards, &commitment);
 //! #
 //! for block in &blocks {
 //!     assert!(verify::<F, G, DP<F>>(block, &powers).unwrap());
@@ -127,7 +127,7 @@
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, prove};
+//! # use komodo::semi_avid::{build, commit};
 //! #
 //! # fn main() {
 //! # let mut rng = ark_std::test_rng();
@@ -139,8 +139,8 @@
 //! #
[…]
