+++ examples/semi_avid.rs
@@ -10,7 +10,7 @@ use komodo::{
     algebra::linalg::Matrix,
     error::KomodoError,
     fec::{decode, encode},
-    semi_avid::{build, prove, recode, verify, Block},
+    semi_avid::{build, commit, recode, verify, Block},
     zk::setup,
 };
 
@@ -37,8 +37,8 @@ where
     eprint!("building blocks... ");
     let encoding_mat = &Matrix::random(k, n, &mut rng);
     let shards = encode(&bytes, encoding_mat)?;
-    let proof = prove(&bytes, &powers, encoding_mat.height)?;
-    let blocks = build::<F, G, P>(&shards, &proof);
+    let commitment = commit(&bytes, &powers, encoding_mat.height)?;
+    let blocks = build::<F, G, P>(&shards, &commitment);
     eprintln!("done");
 
     // verify that all the blocks are valid
