### 2025-12-15 af83d3e
fix benchmarks results branch name (dragoon/komodo!236)
see https://gitlab.isae-supaero.fr/dragoon/komodo/-/snippets/12
