### 2025-12-15 81bac12
remove redundant clones (dragoon/komodo!237)
> [!important]
> this will supersede !217 because we do not have write access to the [`a.bretagne/komodo:main`](https://gitlab.isae-supaero.fr/a.bretagne/komodo/-/tree/main) branch anymore...

performance evaluation will be performed later.

## summary
> [!tip]
> ```bash
> git rev-list main^..HEAD
>     | lines
>     | reverse
>     | each {
>         git checkout $in
>         {
>             rev: $in,
>             clones: (rg '\.clone' **/*.rs | lines | length),
>             to_vecs: (rg '\.to_vec' **/*.rs | lines | length),
>         }
>     }

[…]
