### 2025-12-17 98a96f9
structure the benchmarks (dragoon/komodo!239)
> [!note]
> this is a continuation of b3dcae44 (!233)

## changelog
- use following snippet to run the benchmarks
```bash
const BN254 = { name: "bn254", bits: 254, bytes_without_truncation: 31 }

(make
    (make cartesian-product ...[
            ["semi-avid", "kzg", "aplonk", "fri"]
            (1..10 | each { 2 ** $in })
            (3..18 | each { 2 ** $in * $BN254.bytes_without_truncation })
        ]
        | each {{ p: $in.0, k: $in.1, b: $in.2 }}
        | insert n { 2 * $in.k }
    )
    # --email
    # --shutdown
    --commit
    --seed 1
[…]
