+++ benchmarks/src/random.rs
@@ -1,5 +1,5 @@
-use rand::{rngs::ThreadRng, Rng};
+use rand::Rng;
 
-pub(crate) fn random_bytes(n: usize, rng: &mut ThreadRng) -> Vec<u8> {
+pub(crate) fn random_bytes(n: usize, rng: &mut impl Rng) -> Vec<u8> {
     (0..n).map(|_| rng.gen::<u8>()).collect()
 }
