+++ benchmarks/src/semi_avid.rs
@@ -5,48 +5,90 @@ use ark_ff::PrimeField;
 use ark_poly::DenseUVPolynomial;
 use ark_std::ops::Div;
 
-use rand::thread_rng;
+use rand::Rng;
 
 use komodo::{algebra::linalg::Matrix, fec, semi_avid, zk};
 
-use crate::{name_some_pair, FECParams};
+use crate::FECParams;
+
+pub(crate) struct SemiAVIDResult {
+    t_commit_k: Option<Duration>,
+    t_build_n: Option<Duration>,
+    t_verify_n: Option<Duration>,
+}
+
+impl From<SemiAVIDResult> for Vec<(&'static str, Option<u128>)> {
+    fn from(value: SemiAVIDResult) -> Self {
+        vec![
+            ("t_commit_k", value.t_commit_k.map(|v| v.as_nanos())),
+            ("t_build_n", value.t_build_n.map(|v| v.as_nanos())),
+            ("t_verify_n", value.t_verify_n.map(|v| v.as_nanos())),
+        ]
+    }
+}
+
+impl SemiAVIDResult {
+    /// Sets all fields to [`None`].
+    fn empty() -> Self {
+        SemiAVIDResult {
+            t_commit_k: None,
+            t_build_n: None,
+            t_verify_n: None,
+        }
+    }
+}
 
 pub(crate) fn bench<F, G, P>(
     nb_bytes: usize,
     fec_params: FECParams,
-) -> Vec<(&'static str, Option<Duration>)>
+    rng: &mut impl Rng,
+) -> SemiAVIDResult
 where
     F: PrimeField,
     G: CurveGroup<ScalarField = F>,
     P: DenseUVPolynomial<F>,
     for<'a, 'b> &'a P: Div<&'b P, Output = P>,
 {
-    let mut rng = thread_rng();
-    let bytes = crate::random::random_bytes(nb_bytes, &mut rng);
+    let bytes = crate::random::random_bytes(nb_bytes, rng);
 
-    let powers = zk::setup::<F, G>(zk::nb_elements_in_setup::<F>(bytes.len()), &mut rng).unwrap();
-    let encoding_mat = Matrix::random(fec_params.k, fec_params.n, &mut rng);
-    let shards = fec::encode(&bytes, &encoding_mat).unwrap();
+    let powers = zk::setup::<F, G>(zk::nb_elements_in_setup::<F>(bytes.len()), rng)
+        .expect("komodo::zk::setup");
+    let encoding_mat = Matrix::random(fec_params.k, fec_params.n, rng);
+    let shards = fec::encode(&bytes, &encoding_mat).expect("komodo::fec::encode");
 
     let plnk::TimeWithValue {
         t: t_commit_k,
         v: commitment,
-    } = plnk::timeit(|| semi_avid::commit(&bytes, &powers, encoding_mat.height).unwrap());
+    } = plnk::timeit(|| {
+        semi_avid::commit(&bytes, &powers, encoding_mat.height).expect("komodo::semi_avid::prove")
+    });
 
     let plnk::TimeWithValue {
         t: t_build_n,
         v: blocks,
     } = plnk::timeit(|| semi_avid::build(&shards, &commitment));
 
-    let plnk::TimeWithValue { t: t_verify_n, .. } = plnk::timeit(|| {
-        for block in &blocks {
-            assert!(semi_avid::verify(block, &powers).unwrap());
+    let plnk::TimeWithValue {
+        t: t_verify_n,
+        v: ok,
+    } = plnk::timeit(|| {
+        let mut ok = true;
+        for (i, block) in blocks.iter().enumerate() {
+            if !semi_avid::verify(block, &powers)
[…]
