+++ benchmarks/heat_map.py
@@ -5,34 +5,143 @@
 # ]
 # ///
 import numpy as np
+
 import matplotlib.pyplot as plt
+import matplotlib.cm as cm
+import matplotlib.colors as mcolors
 
 import argparse
+import json
+
+
+class ParseFloatStrPair(argparse.Action):
+    def __init__(self, option_strings, dest, nargs=None, **kwargs):
+        super().__init__(option_strings, dest, nargs=nargs, **kwargs)
+
+    def __call__(self, parser, namespace, values, option_string=None):
+        pairs = []
+        for value in values:
+            tokens = value.split(':')
+            if len(tokens) != 2:
+                parser.error(f"{self.dest}: invalid element '{value}', expected `float:str` pair")
+            else:
+                tick, label = tokens
+                pairs.append((float(tick), label))
+        setattr(namespace, self.dest, pairs)
+
+
+class ParseColorMap(argparse.Action):
+    def __init__(self, option_strings, dest, nargs=None, **kwargs):
+        super().__init__(option_strings, dest, nargs=nargs, **kwargs)
+
+    def __call__(self, parser, namespace, values, option_string=None):
+        if values.startswith("["):
+            cmap = mcolors.ListedColormap(np.array(json.loads(values)) / 255)
+        else:
+            cmap = plt.get_cmap(name=values)
+
+        setattr(namespace, self.dest, cmap)
 
 
 parser = argparse.ArgumentParser()
-parser.add_argument("values"         , nargs="+" , type=float                 )
-parser.add_argument("--width" , "-W"             , type=int   , required=True )
-parser.add_argument("--height", "-H"             , type=int   , required=True )
-parser.add_argument("--figsize"      , nargs=2   , type=int   , default=(16,9))
-parser.add_argument("--dpi"                      , type=int   , default=300   )
-parser.add_argument("--save"                     , type=str                   )
+parser.add_argument("values"         , nargs="*" , type=float                   )
+parser.add_argument("--width" , "-W"             , type=int   , required=True   )
+parser.add_argument("--height", "-H"             , type=int   , required=True   )
+parser.add_argument("--figsize"      , nargs=2   , type=float                   )
+parser.add_argument("--dpi"                      , type=int   , default=300     )
+parser.add_argument("--cbar"                     , action="store_true"          )
+parser.add_argument("--cbarmin"                  , type=float                   )
+parser.add_argument("--cbarmax"                  , type=float                   )
+parser.add_argument("--cbarticks"    , nargs="*" , action=ParseFloatStrPair     )
+parser.add_argument("--cbarlabel"                , type=str                     )
+parser.add_argument("--cbardir"                  , type=str   , choices=["vertical", "horizontal"], default="vertical")
+parser.add_argument("--cmap"                     , action=ParseColorMap                           , default="plasma")
+parser.add_argument("--clut"                     , type=int                     )
+parser.add_argument("--xlabel"                   , type=str                     )
+parser.add_argument("--xticks"       , nargs="*" , action=ParseFloatStrPair     )
+parser.add_argument("--ylabel"                   , type=str                     )
+parser.add_argument("--yticks"       , nargs="*" , action=ParseFloatStrPair     )
+parser.add_argument("--title"                    , type=str                     )
+parser.add_argument("--save"                     , type=str                     )
[…]
