+++ benchmarks/src/macros.rs
@@ -5,3 +5,11 @@ macro_rules! name_some_pair {
         (stringify!($n), Some($n))
     };
 }
+
+/// `name_none_pair!(foo)` -> `("foo", None)`
+#[macro_export]
+macro_rules! name_none_pair {
+    ($n:ident) => {
+        (stringify!($n), None)
+    };
+}
