### 2026-01-05 f804996
benchmarks: campaign 1 (dragoon/komodo!240)
this MR adds measurements on the following sets of parameters + some other changes to the benchmarks scripts

- elliptic curve: $\text{BN254}$ ($\phi_{\text{BN254}} = 31 \text{B}$)
- seed ($s$): $176597932200 + i$ for $i \in \{0..9\}$
- CPU: $\texttt{ee672bb315ea00fe5815f0e20db6aa88017c1ba8355794f411c10a6057377e57}$
- Git: 98a96f9bb1a025c566b0d16e89990d996ab4df75
- protocol ($p$): one of `semi-avid`, `kzg`, `aplonk`, `fri`
- $\log_2 k$: $\quad 1 \rightarrow 10$
- $\rho = 0.5$
- $\log_2 \frac{b}{\phi_{\text{BN254}}}$: $\quad 3 \rightarrow 18$

---

- 
[…]
