+++ benchmarks/src/main.rs
@@ -1,7 +1,7 @@
 use ark_ec::pairing::Pairing;
 use ark_poly::univariate::DensePolynomial;
 
-use clap::{command, Parser};
+use clap::{Parser, Subcommand, ValueEnum};
 use dragoonfri::algorithms::Sha3_256;
 use dragoonfri::dynamic_folding_factor;
 use rand::{RngCore, SeedableRng};
@@ -71,88 +71,124 @@ enum Curve {
 #[derive(Parser)]
 #[command(version, about, long_about = None)]
 struct Cli {
-    #[arg(long)]
-    nb_bytes: usize,
-    #[arg(short)]
-    k: usize,
-    #[arg(short)]
-    n: usize,
+    #[command(subcommand)]
+    command: Option<Commands>,
+}
 
-    #[arg(long)]
-    seed: Option<u64>,
+#[derive(Subcommand)]
+enum Commands {
+    List,
+    Run {
+        #[arg(long)]
+        nb_bytes: usize,
+        #[arg(short)]
+        k: usize,
+        #[arg(short)]
+        n: usize,
 
-    #[arg(long)]
-    protocol: Protocol,
+        #[arg(long)]
+        seed: Option<u64>,
 
-    #[arg(long)]
-    curve: Curve,
+        #[arg(long)]
+        protocol: Protocol,
 
-    #[arg(long)]
-    fri_ff: Option<usize>,
-    #[arg(long)]
-    fri_bf: Option<usize>,
-    #[arg(long)]
-    fri_rpo: Option<usize>,
-    #[arg(long)]
-    fri_q: Option<usize>,
+        #[arg(long)]
+        curve: Curve,
+
+        #[arg(long)]
+        fri_ff: Option<usize>,
+        #[arg(long)]
+        fri_bf: Option<usize>,
+        #[arg(long)]
+        fri_rpo: Option<usize>,
+        #[arg(long)]
+        fri_q: Option<usize>,
+    },
 }
 
 fn main() {
     let args = Cli::parse();
 
-    let fec_params = FECParams {
-        k: args.k,
-        n: args.n,
-    };
+    if let Some(cmd) = args.command {
+        match cmd {
+            Commands::List => {
+                println!(
+                    "protocols:{}",
+                    Protocol::value_variants()
+                        .iter()
+                        .map(|p| p.to_possible_value().unwrap().get_name().to_string())
+                        .collect::<Vec<_>>()
+                        .join(",")
+                );
+                println!(
+                    "curves:{}",
+                    Curve::value_variants()
+                        .iter()
+                        .map(|p| p.to_possible_value().unwrap().get_name().to_string())
+                        .collect::<Vec<_>>()
+                        .join(",")
+                );
+            }
+            Commands::Run {
+                nb_bytes,
+                k,
+                n,
+                seed,
+                protocol,
+                curve,
+                fri_ff,
+                fri_bf,
+                fri_rpo,
+                fri_q,
+            } => {
+                let fec_params = FECParams { k, n };
 
-    let mut rng: Box<dyn RngCore> = match args.seed {
-        Some(seed) => Box::new(rand::rngs::StdRng::seed_from_u64(seed)),
-        None => Box::new(rand::thread_rng()),
-    };
+                let mut rng: Box<dyn RngCore> = match seed {
+                    Some(seed) => Box::new(rand::rngs::StdRng::seed_from_u64(seed)),
+                    None => Box::new(rand::thread_rng()),
+                };
 
-    let kv_pairs: Vec<(&'static str, Option<u128>)> = match args.protocol {
-        Protocol::SemiAVID => match args.curve {
[…]
