### 2026-01-05 5cc8fcc
Remove blocks (dragoon/komodo!238)
> [!note]
> this is related to
> - https://gitlab.isae-supaero.fr/dragoon/komodo/-/issues/22
> - https://gitlab.isae-supaero.fr/dragoon/komodo/-/issues/23

the rationale is that the commit should NOT be part of any kind of _block_ because **(1)** it makes it easy to forge a fake shard and forget to check its validity and **(2)** it is cumbersome with extra functions like `semi_avid::build` and `semi_avid::recode` when `fec` already provides all these functionalities.

with this MR, i propose to **(1)** remove all the `Block` structures, **(2)** cleanup the AP
[…]
