+++ src/semi_avid.rs
@@ -76,12 +76,13 @@
 //! let powers = komodo::zk::setup::<F, G>(10 * 1_024, &mut rng).unwrap();
 //! # }
 //! ```
-//! - we can now build an encoding matrix, encode the data, commit it, proving the shards implicitely, and build [`Block`]s
+//! - we can now build an encoding matrix, encode the data and commit it, proving the shards implicitely
 //! ```
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, commit, verify};
+//! # use komodo::semi_avid::{commit, verify};
+//! # use komodo::algebra::linalg::Matrix;
 //! #
 //! # fn main() {
 //! # let mut rng = ark_std::test_rng();
@@ -91,18 +92,18 @@
 //! #
 //! # let powers = komodo::zk::setup::<F, G>(10 * 1_024, &mut rng).unwrap();
 //! #
-//! let encoding_mat = &komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
-//! let shards = komodo::fec::encode(&bytes, encoding_mat).unwrap();
+//! let encoding_mat: Matrix<F> = komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
+//! let shards = komodo::fec::encode(&bytes, &encoding_mat).unwrap();
 //! let commitment = commit::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
-//! let blocks = build::<F, G, DP<F>>(&shards, &commitment);
 //! # }
 //! ```
-//! - finally, each [`Block`] can be verified individually, using the same trusted setup
+//! - finally, each [`Shard`] can be verified individually, using the same trusted setup
 //! ```
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, commit, verify};
+//! # use komodo::semi_avid::{commit, verify};
+//! # use komodo::algebra::linalg::Matrix;
 //! #
 //! # fn main() {
 //! # let mut rng = ark_std::test_rng();
@@ -112,13 +113,12 @@
 //! #
 //! # let powers = komodo::zk::setup::<F, G>(10 * 1_024, &mut rng).unwrap();
 //! #
-//! # let encoding_mat = &komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
-//! # let shards = komodo::fec::encode(&bytes, encoding_mat).unwrap();
+//! # let encoding_mat: Matrix<F> = komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
+//! # let shards = komodo::fec::encode(&bytes, &encoding_mat).unwrap();
 //! # let commitment = commit::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
-//! # let blocks = build::<F, G, DP<F>>(&shards, &commitment);
 //! #
-//! for block in &blocks {
-//!     assert!(verify::<F, G, DP<F>>(block, &powers).unwrap());
+//! for shard in &shards {
+//!     assert!(verify::<F, G, DP<F>>(shard, &commitment, &powers).unwrap());
 //! }
 //! # }
 //! ```
@@ -127,7 +127,8 @@
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, commit};
[…]

[…]
