### 2026-01-05 #9577 @a-stevan
@danielskatz
hi, back from the holidays 👋

after searching for ["_Rust_" and "_Cryptography_"](https://reviewers.joss.theoj.org/reviewers/search?language=31&area_id=7&keywords=cryptography&name=&commit=Update+search) and then ["_Rust_" and "_Cybersecurity_"](https://reviewers.joss.theoj.org/reviewers/search?language=31&area_id=7&keywords=+cybersecurity&name=&commit=Update+search), we saw the following five potential reviewers !
- [i64](https://github.com/i64)
- [cassiersg](https://github.com/cassiersg)
- [octetsplicer](https://github.com/octetsplicer)
- [nightlark](https://github.com/nightlark)
- [manish](https://github.com/manish)
