### 2026-01-11 #9577 @cassiersg
@danielskatz I can review this paper.
