### 2026-01-16 #9791 @editorialbot
[REVIEW]: Komodo: Cryptographically-proven Erasure Coding
**Submitting author:** @a-stevan (<a href="http://orcid.org/0009-0003-5684-5862">Antoine Stevan</a>)
**Repository:** https://github.com/dragoon-rs/komodo
**Branch with paper.md** (empty if default branch): paper-joss
**Version:** v1.0.1
