### 2026-01-28 #9791 @danielskatz
@cassiersg & @ankitabanerjee015 - it's now been about two weeks, and JOSS's expectation is that the review will progress interatively, ideally finishing in four weeks. In order to get started, as mentioned above, please use the command `@editorialbot generate my checklist` to create your review checklist. @editorialbot commands need to be the first thing in a new comment.
