### 2026-01-29 #9791 @cassiersg
@a-stevan I have a hard time understanding the high-level features and security guarantees of the system.

As far as I understand, it seems that the various schemes can be roughly abstracted as:
1. `encode`: takes input `data` and generates an array of n `share`
2. `gen_blocks` (includes commitment and proof computation): takes all the `share`s and outputs an array of n `block`
3. `verify`: takes a block and verifies the proof (pass/fail output)
4. `extract`: recover a `share` from a `block`
5. `decode`: recover `data` from k `share`.

Is that correct? It seems that there is sometimes an additional setup phase, right?

What is the security guarantee of the scheme?

As a separate question, I have tried building the documentation (`cargo doc` at the root of the repo, on the `paper-joss` branch, rust 1.93.0), and it fails with the following error:
```
error: error reading `katex.html`: No such file or directory (os error 2)

error: `ExternalHtml::load` failed

error: could not document `unicode-ident`
warning: build failed, waiting for other jobs to finish...
error: could not document `typenum`
```
