### 2026-01-30 #9791 @ankitabanerjee015
@a-stevan
Can we include a separate section titled "State of the field" in the software paper?  Please let me know your thoughts, as I could see a review checklist stating
State of the field: Do the authors describe how this software compares with other commonly used packages?
I do see the content as part of Statement of Need, but i think better to keep the 2 sections separate.
