### 2026-02-05 #9791 @a-stevan
@danielskatz @cassiersg @ankitabanerjee015

sorry for the delay, it was holiday time for me last week, so i did not see the discussion...

we have been thinking about answers during the week, they should come tomorrow.
