### 2026-02-05 fa3bbab
fix documentation with/without KaTeX
