### 2026-02-05 #9791 @a-stevan
@cassiersg

> As a separate question, I have tried building the documentation (`cargo doc` at the root of the repo, on the `paper-joss` branch, rust 1.93.0), and it fails with the following error:
>
> ```
> error: error reading `katex.html`: No such file or directory (os error 2)
>
> error: `ExternalHtml::load` failed
>
> error: could not document `unicode-ident`
> warning: build failed, waiting for other jobs to finish...
> error: could not document `typenum`
> ```
to help you _see_ the documentation, this should be fixed with [`fa3bbab`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210/diffs?commit_id=fa3bbabcb3490425cf96f1d2780e7adc91384d81) (see [the new instructions](https://gitlab.isae-supaero.fr/dragoon/komodo/-/tree/paper-joss#the-library)) !

> [!note]
> also, there is a Rust _toolchain_ in Komodo, so it's quite strange that your Cargo picks up version **1.93.0** 🤔
> `cargo --version` at the root gives me **cargo 1.89.0 (c24e10642 2025-06-23)** as per `rust-toolchain.toml` 🤔
