### 2026-02-05 #9791 @tanguy-perennou
> I have a hard time understanding the high-level features and security guarantees of the system.
>
> As far as I understand, it seems that the various schemes can be roughly abstracted as:
>
>     1. `encode`: takes input `data` and generates an array of n `share`
>     2. `gen_blocks` (includes commitment and proof computation): takes all the `share`s and outputs an array of n `block`
>     3. `verify`: takes a block and verifies the proof (pass/fail output)
>     4. `extract`: recover a `share` from a `block`
>     5. `decode`: recover `data` from k `share`.
>
> Is that correct? It seems that there is sometimes an additional setup phase, right?
>
> What is the security guarantee of the scheme?

@cassiersg As suggested in the review process, we will address the above part of your review in two separate issues https://github.com/dragoon-rs/komodo/issues/3 and https://github.com/dragoon-rs/komodo/issues/4
