### 2026-02-10 #9791 @cassiersg
Taking a look back at the [scope of JOSS](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software), I am wondering if the submission qualifies. As far as I can see, the submitted library seems to have been written first as a research artifact: it is the result of research, but it is not clear if it finds application in research.

Indeed, the main goal of the research seems to be the identification, implementation and benchmarking of polynomial commitment schemes in the context of data sharding. I have not seen an real-world use of the library, and the proposed use-cases are integration of the library into distributed systems (so not a "scientific computing"/"solving modeling problem" application).

@a-stevan Did I miss any place where the library is already in use? (That would be interesting to know, independently of JOSS's scope policy.)

cc @danielskatz
