### 2026-02-11 #9791 @danielskatz
@cassiersg - because this was submitted and started before JOSS's change of scope definition, you should see https://joss.readthedocs.io/en/pre-2026-submission/submitting.html#what-we-mean-by-research-software for JOSS's definitions at that time, which is what this is being reviewed under.
