### 2026-02-12 #9791 @a-stevan
@cassiersg @danielskatz

> Taking a look back at the [scope of JOSS](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software), I am wondering if the submission qualifies. As far as I can see, the submitted library seems to have been written first as a research artifact: it is the result of research, but it is not clear if it finds application in research.
>
> Indeed, the main goal of the research seems to be the identification, implementation and benchmarking of polynomial commitment schemes in the context of data sharding. I have not seen an real-world use of the library, and the proposed use-cases are integration of the library into distributed systems (so not a "scientific computing"/"solving modeling problem" application).
>
looking at both the [old scope](https://joss.readthedocs.io/en/pre-2026-submission/submitting.html#what-we-mean-by-research-software) and the [new scope](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) of JOSS, it looks to me that the definition of what a _research software_ is has not changed, i.e. it is a piece of code that
- [ ] solves complex modeling problems in a scientific context
- [ ] supports the functioning of research instruments or the execution of research experiments
- [ ] extracts knowledge from large data sets
- [x] offers a mathematical library, or similar

i think Komodo satisfies the last point by providing a Rust library that can (**1**) encode and decode arbitrary data with redundancy and (**2**) prove the validity of any shard of data independently.

> [@a-stevan](https://github.com/a-stevan) Did I miss any place where the library is already in use? (That would be interesting to know, independently of JOSS's scope policy.)
>
we are not aware of any use of Komodo outside ISAE yet.

however, we currently use Komodo in a few projects internally
- study of the _recoding_ mechanism and the linear dependencies it introduces ([gitlab.isae-supaero.fr:dragoon/nc-diversity](https://gitlab.isae-supaero.fr/dragoon/nc-diversity))
- integration of the library into an experimental real distributed network during an internship ([gitlab.isae-supaero.fr:dragoon/dragoonfly](https://gitlab.isae-supaero.fr/dragoon/dragoonfly))
- continuation of [gitlab.isae-supaero.fr:dragoon/dragoonfly](https://gitlab.isae-supaero.fr/dragoon/dragoonfly) in another internship with a new architecture ([gitlab.isae-supaero.fr:dragoon/dragoonfly](https://gitlab.isae-supaero.fr/dragoon/dragoonfly))

> [!note]
> some of these projects are still private for now because they're just ongoing research ideas

maybe Komodo currently lacks some visibility due to its hosting on [gitlab.isae-supaero.fr](https://gitlab.isae-supaero.fr) (mirroring [gitlab.isae-supaero.fr:dragoon/komodo](https://gitlab.isae-supaero.fr/dragoon/komodo) to [github.com:dragoon-rs/komodo](https://github.com/dragoon-rs/komodo) has not been advertised in any way). If we make it to a publication in JOSS, we hope to see it used by other projects.
