### 2026-02-25 #9791 @cassiersg
After going through the whole review checklist and following discussions here and on the submission's issue tracker, here is my current assessment of the submission.

##
Detailed comments:

+ Missing dependencies: while the core library is easily built using cargo, running benchmarks depends on nushell, which should be documented as it is not a "standard" system dependency. (Also, I would generally recommend writing such scripts in a language that does not require the user to install additional software, especially a shell not available in many distributions.)
+ Community guidelines: the project does not indicate openness to community: there are no explicitly contributing/reporting/discussion venues, the github repo is only a mirror, and the upstream gitlab is not open for registration.
+ Distribution and documentation: I would recommend adding a mention in the readme that the library is available on crates.io and putting a link to docs.rs.
+ Line 79 of the paper seems orphaned from a paragraph or caption.
+ The paper describes 5 operations, but benchmarks only 3 of them, the missing ones should be added.
+ Statement of need / functionality and documentation: This is the part where I'm struggling.
    1. As far as I understand, the library implements 3 "secure sharding" schemes that should be functionally equivalent (same functionality and same security properties), but have different performance characteristics. However, the API is not the same for all the schemes (and therefore, does not always match the high-level explanations). I think this high-level functionality is a crucial missing part from the documentation. I would also strongly suggest that for this high-level functionality to be reflected in the library API, e.g. by means of a single `trait` implemented for the 3 schemes. This high-level API should also make it clear what are the inputs and outputs of each function (something which is not always clear in the paper summary, readme or top-level documentation).
    2. There should be a clear security property stated in the documentation with respect to this API, with careful design of the errors returned when integrity (or other checks) are violated. From [this discussion](https://github.com/dragoon-rs/komodo/issues/3), it seems to me that the examples/API currently omit the a crucial check in the decoding: that the commitment for all the shards is the same. As far as I understand, in the [2023 paper](https://link.springer.com/chapter/10.1007/978-3-031-89518-0_13), the authors argue that a simple Merkle tree commitment is not sufficient, due to not guaranteeing that any set of t shards decode to the same value. This additional property of the proposed schemes should be clearly stated in the documentation (along with the assumptions - this means the initial sharing/proof operation is not trusted), and there should be an example where this property is relevant. Trusted setup assumptions should also be clarified.
    3. The documentation and statement of need in the paper/readme does not provide any clear application example where the security properties of the schemes are needed. For some examples (information dispersal or drone swarms -- the latter is so high-level that it is undescriptive: in which protocol do we need the sharding?), it is not clear why the initial sharing cannot be trusted. The only well-defined example is a reference to Ethereum's proto-Danksharding, which is a complex blockchain technology whose implementation goes way beyond what is implemented in komodo.

## Summary

* I think that this library needs a significant amount of work on its API and documentation to be re-usable.
* There is no clear application for this library. Ideally, there should be a public project where the library is used. There should at least be concrete example problems that this library solves (and for which it is a good solution).
