### 2026-02-25 #9791 @danielskatz
Thanks @cassiersg - this is a nice detailed summary.

@a-stevan - Can you respond to this?

@ankitabanerjee015 - do you generally agree? Do you have any additional points to add?
