### 2026-03-02 69b9656
make lib.rs documentation more clear
