### 2026-03-02 e1ecce2
add Nushell as a clear dependency
