### 2026-03-02 867ebaf
add a note about contributing to the library
