### 2026-03-03 8376c20
merge figure analysis into 1 paragraph
