### 2026-03-03 bac0e1c
add note about omitting encode and decode
