### 2026-03-03 d27f37d
make "benchmarking API" sentence more clear
