### 2026-03-03 677c197
add "what does <protocol> prove?" sections
