### 2026-03-03 1ea6e1c
add a note about Merkle trees
