### 2026-03-03 b081b54
clarify validity
