### 2026-03-03 #9791 @a-stevan
> * Missing dependencies: while the core library is easily built using cargo, running benchmarks depends on nushell, which should be documented as it is not a "standard" system dependency. (Also, I would generally recommend writing such scripts in a language that does not require the user to install additional software, especially a shell not available in many distributions.)
>
I've added a note to the README of the benchmarks in [`e1ecce2ea`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/e1ecce2ea), linking to pre-built binaries for any major release for conveniency.

Unfortunately, we have chosen to not use a POSIX shell for Komodo and most of the DRAGOON projects, so it will likely stay written in Nushell for the foreseable future.

> * Community guidelines: the project does not indicate openness to community: there are no explicitly contributing/reporting/discussion venues, the github repo is only a mirror, and the upstream gitlab is not open for registration.
>
There are some contacts and links at the very end of [the paper] but a section "contributing" has been added to the main README in [`867ebaf62`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/867ebaf62).

> * Distribution and documentation: I would recommend adding a mention in the readme that the library is available on crates.io and putting a link to docs.rs.
>
Because the [latest release of Komodo on crates.io](https://crates.io/crates/komodo/versions) is already obsolete and behind the current GitLab project, we are thinking about removing Komodo from crates.io and not supporting it in the future on this platform.

> * Line 79 of the paper seems orphaned from a paragraph or caption.
>
Lines 76-79 are indeed the same paragraph, they have been merged in [`8376c209b`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/8376c209b) for readability.

> * The paper describes 5 operations, but benchmarks only 3 of them, the missing ones should be added.
>
The `encode` and `decode` steps are the same for all the methods, i.e. FEC encodings and decodings through matrix multiplications and inversions.

They have been measured in https://gitlab-pages.isae-supaero.fr/dragoon/komodo-benchmark-results/.

A note has been added to the paper in [`bac0e1cb0`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/bac0e1cb0) with indicative values for `encode` and `decode`.

> * Statement of need / functionality and documentation:
>
> 1. As far as I understand, the library implements 3 "secure sharding" schemes that should be functionally equivalent (same functionality and same security properties), but have different performance characteristics. However, the API is not the same for all the schemes (and therefore, does not always match the high-level explanations). I think this high-level functionality is a crucial missing part from the documentation. I would also strongly suggest that for this high-level functionality to be reflected in the library API, e.g. by means of a single `trait` implemented for the 3 schemes. This high-level API should also make it clear what are the inputs and outputs of each function (something which is not always clear in the paper summary, readme or top-level documentation).
>
We have tried a common _trait_ in the past (see [common trait]) but, because the methods require different arguments in all stages and thus introduce optional artificial parameters (that also need to be checked at runtime and potentially throw ad-hoc errors), we decided to abandon the idea.

There was a confusion in [the paper] on line 63 that should have been addressed in [`d27f37d1e`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/d27f37d1e).
The unified API is in the benchmarks because, from the point of view of the benchmarks, the free parameters are the size of the data, $k$ and $n$, which are the same for all protocols.

However, the internal API of Komodo, even though high-level because it abstracts a lot of computation and data manipulation, is not unified.

> * Statement of need / functionality and documentation:
>
> 2. There should be a clear security property stated in the documentation with respect to this API, with careful design of the errors returned when integrity (or other checks) are violated. From [this discussion](https://github.com/dragoon-rs/komodo/issues/3), it seems to me that the examples/API currently omit the a crucial check in the decoding: that the commitment for all the shards is the same. As far as I understand, in the [2023 paper](https://link.springer.com/chapter/10.1007/978-3-031-89518-0_13), the authors argue that a simple Merkle tree commitment is not sufficient, due to not guaranteeing that any set of t shards decode to the same value. This additional property of the proposed schemes should be clearly stated in the documentation (along with the assumptions - this means the initial sharing/proof operation is not trusted), and there should be an example where this property is relevant. Trusted setup assumptions should also be clarified.
>
i've added 4 commits ([`69b965633`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/69b965633), [`d56cdad1f`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/d56cdad1f), [`677c19743`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/677c19743), [`1ea6e1c65`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/1ea6e1c65)) to hopefully clarify the situation a bit
- the top-level doc now has a section that explains the goal of Komodo and its guarantees
- each protocol's submodule now has a section that should explain the goal of this particular protocol
- i've added a note in the top-level documentation about Merkle trees and why they are not enough
- the FEC documentation has been slightly tweaked

Actual security guarantees/proofs are very technical, the reader is encouraged to have a look at the referenced papers for further details.

> * Statement of need / functionality and documentation:
>
> 3. The documentation and statement of need in the paper/readme does not provide any clear application example where the security properties of the schemes are needed. For some examples (information dispersal or drone swarms -- the latter is so high-level that it is undescriptive: in which protocol do we need the sharding?), it is not clear why the initial sharing cannot be trusted. The only well-defined example is a reference to Ethereum's proto-Danksharding, which is a complex blockchain technology whose implementation goes way beyond what is implemented in komodo.
>
In order to clarify your first two points, we have made changes in [`796646c28`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/796646c28) and [`b081b5402`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/b081b5402), introducing the larger concept of _validity_ instead of _integrity_.

We fully agree that Komodo is largely smaller than Ethereum, and smaller than proto-Danksharding since it does not cover the sampling part.
However, we believe that Komodo is comparable to the primitives of Ethereum proto-Danksharding that encode shards, proves them and verifies them.

---

please let us know if we have missed anything.

[the paper]: https://github.com/openjournals/joss-papers/blob/joss.09791/joss.09791/10.21105.joss.09791.pdf
[common trait]: https://gitlab.isae-supaero.fr/dragoon/pcs-fec-id/-/blob/master/src/lib.rs#L279-305
