### 2026-03-10 #9791 @cassiersg
My main concern regarding the feature (i.e. security guarantees) is still not addressed: the documented property does not provide any relevance to the high-level feature, I do not agree with author's answer on the API, and there is still no concrete use-case.

# On the feature

On the notion of **validity**.

> a given shard   is valid, i.e. has been generated indeed through an expected encoding process,  namely a polynomial evaluation or vector inner-product encoding such as  Reed-Solomon.

this notion is satisfied for any value of the shard: any (vector of) point(s) is the value of some (in general of many) polynomial at a point `x`. Therefore, any share, no matter how it has been generated, is valid according to this definition.

## An alternative definition

In my understanding, the high-level feature is roughly the following: a shard is associated with an index in `0,...,n-1` and a (short) commit. We can verify the validity of the shard with respect to the index and the commit. Further, given a commit, there exists a corresponding data. Given a commit, this data is equal to the decoding any set of t shards that are valid w.r.t. t distinct indices and that commit.

## High-level vs low-level

Most of the documentation is focused on polynomial commitments (and other similar techniques). From reading the paper, it seems to me that the originality of the works lies in building a higher-level primitive: verifiable data sharding.

Therefore, I would expect the library APIs and documentation to be expressed w.r.t. that primitive. However, the library contains a set of low-level (FEC, commitments, ZKP) that can be used to implement the high-level functionality, but the high-level functionality itself is not implemented.

 In particular, the guarantee on the data sharding should be the same irrespective of the commitment&proof system being used (up to the quantitative advantage of the adversary).

# On the API

This follows from the above low-level vs high-level discussion: the [common API attempt](https://gitlab.isae-supaero.fr/dragoon/pcs-fec-id/-/blob/master/src/lib.rs#L279-305) is for the low-level API (polynomial commitment + evaluation proof). This low-level depends indeed significantly on the scheme used (since not all of them are strictly polynomial commitments). However a high-level data sharding API should be possible.

To make it more concrete, here is a very rough idea of what such an API could look like:

```rust
trait VerifiableDataSharding {
    type Params;
    type Commit: Copy;
    fn new(n: u32, t: u32, params: Params) -> Self;
    fn shard(&self, data: &[u8]) -> (Commit, Vec<Shard>);
    fn verify(&self, commit: Commit, i: u32, shard: &Shard) -> bool;
    fn decode(&self, shards: &[(u32, &Shard)]) -> Result<Vec<u8>, IndicesError>;
    fn verified_decode(&self, commit: Commit, shards: &[(u32, &Shard)]) -> Result<Vec<u8>, VDecError> {
        // VDecError is either shard verification or error in the indices.
        // Check all the shares with verify, and if they are all valid, run decode.
    }
}
```

(The fundamental property can then be expressed based on this API: `verified_decode` returns the input data of a `shard` call if it receives `t` shards with the correct indices and commit generated from that call. Otherwise, it returns an error.)

A complete/correct API might look quite different, by my point is that since the paper claims 3 implementations for the same abstract data type, there should be a single API for that data type, and the properties of the data type should be documented in terms of this API.

# On the use-cases

The paper (and documentation) still lacks a relevant and substantiated use-case.
