### 2026-03-11 #9791 @a-stevan
thanks again @cassiersg for your feedback, you are raising interesting points we have overlooked during our internal reviews !

we will be working on this in the following days and will come back to this discussion as soon as possible

the API might take a tiny bit longer
