### 2026-03-11 c1b69e5
reformulate intro and split lines on periods
