### 2026-03-11 4bc333a
switch to github.com:fornwall/rust-script (dragoon/komodo!243)
should address #31 

the CI Docker image has been updated as well: `406e2b0c` -> `31e608d6`.
