### 2026-03-11 97aeea7
rename `semi_avid::prove` into `semi_avid::commit`
