+++ src/semi_avid.rs
@@ -82,12 +82,12 @@
 //! let powers = komodo::zk::setup::<F, G>(10 * 1_024, &mut rng).unwrap();
 //! # }
 //! ```
-//! - we can now build an encoding matrix, encode the data, prove the shards and build [`Block`]s
+//! - we can now build an encoding matrix, encode the data, commit it and build [`Block`]s
 //! ```
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, prove, verify};
+//! # use komodo::semi_avid::{build, commit, verify};
 //! #
 //! # fn main() {
 //! # let mut rng = ark_std::test_rng();
@@ -99,8 +99,8 @@
 //! #
 //! let encoding_mat = &komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
 //! let shards = komodo::fec::encode(&bytes, encoding_mat).unwrap();
-//! let proof = prove::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
-//! let blocks = build::<F, G, DP<F>>(&shards, &proof);
+//! let commitment = commit::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
+//! let blocks = build::<F, G, DP<F>>(&shards, &commitment);
 //! # }
 //! ```
 //! - finally, each [`Block`] can be verified individually, using the same trusted setup
@@ -108,7 +108,7 @@
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, prove, verify};
+//! # use komodo::semi_avid::{build, commit, verify};
 //! #
 //! # fn main() {
 //! # let mut rng = ark_std::test_rng();
@@ -120,8 +120,8 @@
 //! #
 //! # let encoding_mat = &komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
 //! # let shards = komodo::fec::encode(&bytes, encoding_mat).unwrap();
-//! # let proof = prove::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
-//! # let blocks = build::<F, G, DP<F>>(&shards, &proof);
+//! # let commitment = commit::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
+//! # let blocks = build::<F, G, DP<F>>(&shards, &commitment);
 //! #
 //! for block in &blocks {
 //!     assert!(verify::<F, G, DP<F>>(block, &powers).unwrap());
@@ -133,7 +133,7 @@
 //! # use ark_bls12_381::{Fr as F, G1Projective as G};
 //! # use ark_poly::univariate::DensePolynomial as DP;
 //! #
-//! # use komodo::semi_avid::{build, prove};
+//! # use komodo::semi_avid::{build, commit};
 //! #
 //! # fn main() {
 //! # let mut rng = ark_std::test_rng();
@@ -145,8 +145,8 @@
 //! #
 //! # let encoding_mat = &komodo::algebra::linalg::Matrix::random(k, n, &mut rng);
 //! # let shards = komodo::fec::encode(&bytes, encoding_mat).unwrap();
-//! # let proof = prove::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
-//! # let blocks = build::<F, G, DP<F>>(&shards, &proof);
+//! # let commitment = commit::<F, G, DP<F>>(&bytes, &powers, encoding_mat.height).unwrap();
[…]
