### 2026-03-11 5ab7d4e
link to fec module in lib.rs doc
