+++ paper.bib
@@ -102,7 +102,7 @@
   author       = {Stevan, Antoine and Lavaur, Thomas and Lacan, J{\'e}r{\^o}me and Detchart, Jonathan and P{\'e}rennou, Tanguy},
   booktitle    = {International Conference on Information Systems Security and Privacy},
   pages        = {274--300},
-  year         = {2023},
+  year         = {2025},
   doi          = {10.1007/978-3-031-89518-0_13},
   organization = {Springer},
   url          = {https://link.springer.com/chapter/10.1007/978-3-031-89518-0_13},
@@ -115,6 +115,25 @@
     journal = {10th International Conference on Information Systems Security and Privacy},
     year    = {2024},
     doi     = {10.5220/0012377900003648},
+}
+
+@misc{rsmt2d,
+  author    = {{rsmt2d contributors}},
+  title     = {Go implementation of two dimensional Reed-Solomon Merkle tree data availability scheme},
+  year      = {2026},
+  publisher = {GitHub},
+  journal   = {GitHub repository},
+  url       = {https://github.com/celestiaorg/rsmt2d},
+}
+
+@misc{albassam2019frauddataavailabilityproofs,
+      title={Fraud and Data Availability Proofs: Maximising Light Client Security and Scaling Blockchains with Dishonest Majorities},
+      author={Mustafa Al-Bassam and Alberto Sonnino and Vitalik Buterin},
+      year={2019},
+      eprint={1809.09044},
+      archivePrefix={arXiv},
+      primaryClass={cs.CR},
+      url={https://arxiv.org/abs/1809.09044},
 }
 
 @draft{stevan2025komodo,
