### 2026-03-12 3a55b63
add a common `Protocol` trait with doctest
