### 2026-03-12 6dea6e2
make `aplonk::setup` take an RNG
