+++ src/aplonk/mod.rs
@@ -18,7 +18,7 @@ use ark_poly_commit::{
     PCRandomness,
 };
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize, Compress};
-use ark_std::{test_rng, One, UniformRand};
+use ark_std::{rand::Rng, One, UniformRand};
 use rs_merkle::algorithms::Sha256;
 use rs_merkle::Hasher;
 use std::ops::{Div, Mul};
@@ -86,13 +86,13 @@ pub struct SetupParams<E: Pairing> {
 pub fn setup<E, P>(
     degree_bound: usize,
     nb_polynomials: usize,
+    rng: &mut impl Rng,
 ) -> Result<SetupParams<E>, ark_poly_commit::Error>
 where
     E: Pairing,
     P: DenseUVPolynomial<E::ScalarField, Point = E::ScalarField>,
     for<'a, 'b> &'a P: Div<&'b P, Output = P>,
 {
-    let rng = &mut test_rng();
     let params = KZG10::<E, P>::setup(degree_bound, true, rng)?;
 
     let g_1 = params.powers_of_g[0];
@@ -433,6 +433,7 @@ mod tests {
     use ark_ec::{pairing::Pairing, AffineRepr};
     use ark_ff::{Field, PrimeField};
     use ark_poly::{univariate::DensePolynomial, DenseUVPolynomial};
+    use ark_std::{rand::Rng, test_rng};
     use std::ops::{Div, MulAssign};
 
     type UniPoly381 = DensePolynomial<<Bls12_381 as Pairing>::ScalarField>;
@@ -447,6 +448,7 @@ mod tests {
         bytes: &[u8],
         k: usize,
         n: usize,
+        rng: &mut impl Rng,
     ) -> Result<
         (
             super::Commitment<E>,
@@ -464,7 +466,7 @@ mod tests {
         let vector_length_bound =
             bytes.len() / (E::ScalarField::MODULUS_BIT_SIZE as usize / 8) / (degree + 1);
 
-        let params = super::setup::<E, P>(degree, vector_length_bound)?;
+        let params = super::setup::<E, P>(degree, vector_length_bound, rng)?;
         let (_, vk_psi) = trim(&params.kzg, degree);
 
         let elements = algebra::split_data_into_field_elements::<E::ScalarField>(bytes, k);
@@ -497,14 +499,19 @@ mod tests {
         ))
     }
 
-    fn verify_template<E, P>(bytes: &[u8], k: usize, n: usize) -> Result<(), ark_poly_commit::Error>
+    fn verify_template<E, P>(
+        bytes: &[u8],
+        k: usize,
+        n: usize,
+        rng: &mut impl Rng,
+    ) -> Result<(), ark_poly_commit::Error>
     where
         E: Pairing,
         P: DenseUVPolynomial<E::ScalarField, Point = E::ScalarField>,
         for<'a, 'b> &'a P: Div<&'b P, Output = P>,
     {
         let (commitment, blocks, vk) =
-            test_setup::<E, P>(bytes, k, n).expect("proof failed for bls12-381");
+            test_setup::<E, P>(bytes, k, n, rng).expect("proof failed for bls12-381");
 
         for (i, (shard, proof)) in blocks.iter().enumerate() {
             assert!(super::verify::<E, P>(
@@ -524,6 +531,7 @@ mod tests {
         bytes: &[u8],
         k: usize,
         n: usize,
+        rng: &mut impl Rng,
     ) -> Result<(), ark_poly_commit::Error>
     where
         E: Pairing,
@@ -531,7 +539,7 @@ mod tests {
         for<'a, 'b> &'a P: Div<&'b P, Output = P>,
     {
[…]
