### 2026-03-13 dedefcf
Make FEC rererences more consistent
