+++ paper.md
@@ -32,11 +32,11 @@ bibliography: paper.bib
 
 # Summary
 
-We present **Komodo**, a library that allows to (**1**) encode data with erasure-code techniques such as Reed-Solomon encoding, (**2**) prove the resulting _shards_ with cryptographic protocols, (**3**) verify their validity on the other end of any distributed network and (**4**) decode the original data from a subset of said _shards_; [@stevan2024performance] and [@stevan2023assessing].
+We present **Komodo**, a library that allows to (**1**) encode data with forward erasure-code (FEC) techniques such as Reed-Solomon encoding, (**2**) prove the resulting _shards_ with cryptographic protocols, (**3**) verify their validity on the other end of any distributed network and (**4**) decode the original data from a subset of said _shards_; [@stevan2024performance] and [@stevan2023assessing].
 The library is implemented in the _Rust_ programming language and available on the ISAE-SUPAERO GitLab instance [^1] with a mirror on GitHub [^2], both released under the MIT license.
 **Komodo** should be of interest for people willing to explore the field of cryptographically-proven _shards_ of data in distributed systems or data availability sampling settings.
 
-A _shard_ of data is said to be _valid_ if it has been generated indeed through an expected encoding process, namely a polynomial evaluation or vector inner-product encoding such as Reed-Solomon.
+A _shard_ of data is said to be _valid_ if it has been generated by a correct FEC encoding of the data.
 
 **Komodo** implements a protocol API to achieve the following on any input data in a distributed network:
 
