### 2026-03-13 #9791 @a-stevan
Before moving on to the review changes, we have merged pending changes from the `main` branch to the `paper-joss` branch in [4bdd34628](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/4bdd34628).
These were merged into `main` before the review started and mainly focus on internal changes.

> On the notion of **validity**.
>
> > a given shard   is valid, i.e. has been generated indeed through an expected encoding process,  namely a polynomial evaluation or vector inner-product encoding such as  Reed-Solomon.
>
> this notion is satisfied for any value of the shard: any (vector of) point(s) is the value of some (in general of many) polynomial at a point `x`. Therefore, any share, no matter how it has been generated, is valid according to this definition.
>
This is not what we had in mind and tried to explain in the paper and documentation.
The intro of the paper has been reformulated in [c1b69e532](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/c1b69e532) (extracted the definition of _validity_ and split the `commit` and `prove` steps).
The documentation has also been reworked a bit in the previous review iteration:
- added a "_what does komodo prove_" section to the top-level doc
- added individual "_what does the protocol prove_" section to submodules of each protocol

**Please let us know if something remains unclear.**

> ## An alternative definition
>
> In my understanding, the high-level feature is roughly the following: a shard is associated with an index in 0,...,n-1 and a (short) commit. We can verify the validity of the shard with respect to the index and the commit. Further, given a commit, there exists a corresponding data. Given a commit, this data is equal to the decoding any set of t shards that are valid w.r.t. t distinct indices and that commit.
>
We do not have the same definition of a commitment in **Komodo**: a _commitment_ is a kind of fingerprint of the whole original data.
It is common to all encoded shards.

Then an individual proof is attached to each shard separately.

The triplet (commitment, shard, proof) is used to verify the _validity_ of said shard.

Also, the commitment is not used to decode the data, what also explains the separation between the FEC mechanisms and cryptographic protocols in the code base.

> In my understanding, the high-level feature is roughly the following: a shard is associated with an index in `0,...,n-1` and a (short) commit. We can verify the validity of the shard with respect to the index and the commit. Further, given a commit, there exists a corresponding data. Given a commit, this data is equal to the decoding any set of t shards that are valid w.r.t. t distinct indices and that commit.
> ## High-level vs low-level
>
> Most of the documentation is focused on polynomial commitments (and other similar techniques). From reading the paper, it seems to me that the originality of the works lies in building a higher-level primitive: verifiable data sharding.
>
> Therefore, I would expect the library APIs and documentation to be expressed w.r.t. that primitive. However, the library contains a set of low-level (FEC, commitments, ZKP) that can be used to implement the high-level functionality, but the high-level functionality itself is not implemented.
>
> In particular, the guarantee on the data sharding should be the same irrespective of the commitment&proof system being used (up to the quantitative advantage of the adversary).
> # On the API
>
> This follows from the above low-level vs high-level discussion: the [common API attempt](https://gitlab.isae-supaero.fr/dragoon/pcs-fec-id/-/blob/master/src/lib.rs#L279-305) is for the low-level API (polynomial commitment + evaluation proof). This low-level depends indeed significantly on the scheme used (since not all of them are strictly polynomial commitments). However a high-level data sharding API should be possible.
>
> To make it more concrete, here is a very rough idea of what such an API could look like:
>
> trait VerifiableDataSharding {
>     type Params;
>     type Commit: Copy;
>     fn new(n: u32, t: u32, params: Params) -> Self;
>     fn shard(&self, data: &[u8]) -> (Commit, Vec<Shard>);
>     fn verify(&self, commit: Commit, i: u32, shard: &Shard) -> bool;
>     fn decode(&self, shards: &[(u32, &Shard)]) -> Result<Vec<u8>, IndicesError>;
>     fn verified_decode(&self, commit: Commit, shards: &[(u32, &Shard)]) -> Result<Vec<u8>, VDecError> {
>         // VDecError is either shard verification or error in the indices.
>         // Check all the shares with verify, and if they are all valid, run decode.
>     }
> }
>
> (The fundamental property can then be expressed based on this API: `verified_decode` returns the input data of a `shard` call if it receives `t` shards with the correct indices and commit generated from that call. Otherwise, it returns an error.)
>
> A complete/correct API might look quite different, by my point is that since the paper claims 3 implementations for the same abstract data type, there should be a single API for that data type, and the properties of the data type should be documented in terms of this API.
>
Thanks for your suggestion, a common trait has been added in [3a55b63aa](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/3a55b63aa).
It has been implemented for all three protocols in their respective submodules.
The top-level documentation example has been updated to use this trait and features testable snippets of code.
The complete example can be found at `./examples/trait.rs`.

The intent of Komodo, as stated at the very beginning of the documentation, is to provide two parts:
- the encoding / decoding with `fec` module
- separate submodules for the cryptographic protocols

Thus this new trait does not include the `encode` and `decode` steps because these are out of the scope of the cryptographic protocols.

During the creation of the trait, we have identified some issues in the previous internal API of the protocols
- `semi_avid::prove` has been renamed into `semi_avid::commit` ([97aeea78f](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/97aeea78f))
- `aplonk::setup` now takes an RNG as parameter instead of hardcoding a test one ([6dea6e2c4](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/6dea6e2c4))

> # On the use-cases
>
> The paper (and documentation) still lacks a relevant and substantiated use-case.
>
We have updated the _statement of need_ in [1211d6bd8](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/1211d6bd8), specifically the first paragraphs, as an attempt to clarify the use-case of Komodo.

**Is that ok with you ?**
