### 2026-04-01 #9791 @cassiersg
Thanks @a-stevan . Your latest changes addressed my major concerns.

The main question that remains for me is the one of use-cases in the library: even with the additional explanations, whether the library will end up being used (by others that the authors and immediate collaborators) is unclear, and it seems to me that the use-case motivation is weaker than most published JOSS papers. However, I do not see this limitation as necessarily blocking, and defer to others for a decision on this point.

cc @danielskatz
