### 2026-04-02 #9791 @danielskatz
👋 @cassiersg - I see a lot of unchecked items on your checklist. Can you check off the ones you feel are ok?

And given that this was submitted under JOSS's prior scope rules, I don't see the issue you mention (the use-case motivation) as blocking, as long as it is being used for research and could be by others.
