### 2026-04-03 #9791 @ankitabanerjee015
I went over the conversation above and the explanation that @a-stevan  has provided . statement of need looks good to me.
I cross verifies and my questions were addressed. I suggestion is if the below proposal from @tanguy-perennou can we included that will be great. https://github.com/dragoon-rs/komodo/issues/1

We propose to insert a link to the [gitlab.isae-supaero.fr:dragoon/komodo-benchmark-results](https://gitlab.isae-supaero.fr/dragoon/komodo-benchmark-results) repo in paragraph 76-79, in a sentence explaining that dataset is available for anyone. Is it OK for you?
