+++ paper.md
@@ -53,9 +53,9 @@ A _shard_ of data is said to be _valid_ if it has been generated by a correct FE
 
 **Komodo** provides the three following cryptographic protocols:
 
-- **KZG+**: based on [@kate2010constant] and its multi-polynomial extension [@boneh2020efficient]
-- **aPlonK**: based on **PlonK** [@gabizon2019plonk] and **aPlonK** [@ambrona2023aplonk]
-- **Semi-AVID**: based on **Semi-AVID-PR** [@nazirkhanova2022information]
+- **KZG+** (short for the first letters of the authors' names): based on [@kate2010constant] and its multi-polynomial extension [@boneh2020efficient]
+- **aPlonK** (short for _**A**ggregated **P**ermutations over **L**agrange-bases for **O**ecumenical **N**oninteractive arguments of **K**nowledge_): based on **PlonK** [@gabizon2019plonk] and **aPlonK** [@ambrona2023aplonk]
+- **Semi-AVID** (short for _Semi **A**synchronous **V**erifiable **I**nformation **D**ispersal_): based on **Semi-AVID-PR** [@nazirkhanova2022information]
 
 **Komodo** is based on the Arkworks library [@arkworks] which provides implementations of elliptic curves, fields and polynomial algebra.
 
