### 2026-04-07 66b44ad
mention RS codes when introducing "encode"
