+++ paper.md
@@ -40,7 +40,7 @@ A _shard_ of data is said to be _valid_ if it has been generated by a correct FE
 
 **Komodo** implements a protocol API to achieve the following on any input data in a distributed network:
 
-- `encode`: encodes data into _shards_ with a $(k,n)$-code.
+- `encode`: encodes data into _shards_ with a Reed-Solomon $(k,n)$-code.
   This adds redundancy to the data, making the network more resilient to failure, fragmentation, partitioning, loss or corruption.
 - `commit`: commit the input data into a _commitment_ common to all _shards_.
 - `prove`: generate one proof for all $n$ _shards_ with one of three available cryptographic protocols (see below for more information).
