+++ paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 # Summary
 
 We present **Komodo**, a library that allows to (**1**) encode data with forward erasure-code (FEC) techniques such as Reed-Solomon encoding, (**2**) prove the resulting _shards_ with cryptographic protocols, (**3**) verify their validity on the other end of any distributed network and (**4**) decode the original data from a subset of said _shards_; [@stevan2024performance] and [@stevan2023assessing].
-The library is implemented in the _Rust_ programming language and available on the ISAE-SUPAERO GitLab instance [^1] with a mirror on GitHub [^2], both released under the MIT license.
+The library is implemented in the _Rust_ programming language and available on the ISAE-SUPAERO GitLab instance[^1] with a mirror on GitHub[^2], both released under the MIT license.
 **Komodo** should be of interest for people willing to explore the field of cryptographically-proven _shards_ of data in distributed systems or data availability sampling settings.
 
 A _shard_ of data is said to be _valid_ if it has been generated by a correct FEC encoding of the data.
@@ -58,9 +58,6 @@ A _shard_ of data is said to be _valid_ if it has been generated by a correct FE
 - **Semi-AVID** (short for _Semi **A**synchronous **V**erifiable **I**nformation **D**ispersal_): based on **Semi-AVID-PR** [@nazirkhanova2022information]
 
 **Komodo** is based on the Arkworks library [@arkworks] which provides implementations of elliptic curves, fields and polynomial algebra.
-
-[^1]: GitLab source code: [https://gitlab.isae-supaero.fr/dragoon/komodo](https://gitlab.isae-supaero.fr/dragoon/komodo)
-[^2]: GitHub mirror for issues and pull requests: [https://github.com/dragoon-rs/komodo](https://github.com/dragoon-rs/komodo)
 
 # Keywords
 
@@ -125,6 +122,8 @@ been omitted from the analysis below. `encode` involves a matrix multiplication
 which takes around $10\mu s$ with $k = 8$. `decode` has an extra matrix inversion
 of size $k = 8$ taking around $250 \mu s$.
 
+All raw results have been pushed to a separate dedicated repository[^3] and all the pre-compiled figures are available on the _GitLab pages_ instance of ISAE SUPAERO[^4].
+
 ![Performance for small files. Average over $10$ runs.\label{fig:small}](figures/99613e59eb168636525c71d3f3d7a71fa773912ff80fbc70db035d076468633f.png)
 
 ![Performance for large files. Average over $10$ runs for \textbf{KZG+} and \textbf{Semi-AVID}. Only $1$ run for \textbf{aPlonK}. \label{fig:large}](figures/30f6bd95df8c5bd92d9d45585c5050a2e41be2814fd31b0f54f268d9bbbe3d3f.png)
@@ -150,3 +149,8 @@ Ministry of Defense through the Research Project DRAGOON: Dependable distRibuted
 storAGe fOr mObile Nodes (2022 65 0082).
 
 # References
+
+[^1]: GitLab source code: [https://gitlab.isae-supaero.fr/dragoon/komodo](https://gitlab.isae-supaero.fr/dragoon/komodo)
+[^2]: GitHub mirror for issues and pull requests: [https://github.com/dragoon-rs/komodo](https://github.com/dragoon-rs/komodo)
+[^3]: results repository: [https://gitlab.isae-supaero.fr/dragoon/komodo-benchmark-results](https://gitlab.isae-supaero.fr/dragoon/komodo-benchmark-results)
+[^4]: pre-compiled results figures: [https://gitlab-pages.isae-supaero.fr/dragoon/komodo-benchmark-results/](https://gitlab-pages.isae-supaero.fr/dragoon/komodo-benchmark-results/)
