### 2026-04-07 #9791 @a-stevan
👋 @danielskatz

The last points of https://github.com/dragoon-rs/komodo/issues/1 should have been addressed in a few additional commits and https://github.com/dragoon-rs/komodo/issues/2 should be good to go !

As for https://github.com/dragoon-rs/komodo/issues/3, https://github.com/dragoon-rs/komodo/issues/4 and https://github.com/dragoon-rs/komodo/issues/5, they should also have been addressed during the conversation above, if we did not forget about anything of course.

> **EDIT**: i closed the three last issues above because i think the discussion moved to here rather than in the separate issues. They can obviously be reopened !
