### 2026-04-07 #9791 @a-stevan
@cassiersg @ankitabanerjee015

the final version of the PDF so far should be available when the bot is finished !
