### 2026-04-07 #9791 @danielskatz
👋 @cassiersg @ankitabanerjee015 - Can you take another look and see what you now think, and if you can complete your checklists or make more progress on them?
