### 2026-04-23 eb705ff
minor changes in paper.md
