+++ paper.md
@@ -32,8 +32,8 @@ bibliography: paper.bib
 
 # Summary
 
-We present **Komodo**, a library that allows to (**1**) encode data with forward erasure-code (FEC) techniques such as Reed-Solomon encoding, (**2**) prove the resulting _shards_ with cryptographic protocols, (**3**) verify their validity on the other end of any distributed network and (**4**) decode the original data from a subset of said _shards_; [@stevan2024performance] and [@stevan2023assessing].
-The library is implemented in the _Rust_ programming language and available on the ISAE-SUPAERO GitLab instance[^1] with a mirror on GitHub[^2], both released under the MIT license.
+We present **Komodo**, a library that allows one to (**1**) encode data with forward erasure-code (FEC) techniques such as Reed-Solomon encoding, (**2**) prove the resulting _shards_ with cryptographic protocols, (**3**) verify their validity on the other end of any distributed network, and (**4**) decode the original data from a subset of said _shards_ [@stevan2024performance; @stevan2023assessing].
+The library is implemented in the _Rust_ programming language and available under the MIT licenseon the ISAE-SUPAERO GitLab instance[^1] with a mirror on GitHub[^2].
 **Komodo** should be of interest for people willing to explore the field of cryptographically-proven _shards_ of data in distributed systems or data availability sampling settings.
 
 A _shard_ of data is said to be _valid_ if it has been generated by a correct FEC encoding of the data.
@@ -53,26 +53,18 @@ A _shard_ of data is said to be _valid_ if it has been generated by a correct FE
 
 **Komodo** provides the three following cryptographic protocols:
 
-- **KZG+** (short for the first letters of the authors' names): based on [@kate2010constant] and its multi-polynomial extension [@boneh2020efficient]
+- **KZG+** (short for the first letters of the authors' names): based on @kate2010constant and its multi-polynomial extension [@boneh2020efficient]
 - **aPlonK** (short for _**A**ggregated **P**ermutations over **L**agrange-bases for **O**ecumenical **N**oninteractive arguments of **K**nowledge_): based on **PlonK** [@gabizon2019plonk] and **aPlonK** [@ambrona2023aplonk]
 - **Semi-AVID** (short for _Semi **A**synchronous **V**erifiable **I**nformation **D**ispersal_): based on **Semi-AVID-PR** [@nazirkhanova2022information]
 
 **Komodo** is based on the Arkworks library [@arkworks] which provides implementations of elliptic curves, fields and polynomial algebra.
 
-# Keywords
-
-Cryptography;
-Erasure codes;
-Distributed systems;
-Verifiable information dispersal;
-Data availability;
-
 # Statement of need
 
-The verification mechanisms provided by **Komodo** can help implementing verifiable information
+The verification mechanisms provided by **Komodo** can help in implementing verifiable information
 dispersal or data availability schemes in a distributed storage system or in a blockchain.
 
-For instance, in a distributed storage system, nodes can use **Komodo** to encode an incoming data
+For instance, in a distributed storage system, nodes can use **Komodo** to encode incoming data
 into shards; commit them to the initial data; compute individual shard proofs; bundle shards,
 commitments and proofs into messages; and distribute these messages across the network. Any node that
 receives a block can then verify its validity before storing or forwarding it. This ensures data
@@ -90,7 +82,7 @@ A few libraries provide similar functionalities, with a few gaps filled by **Kom
 
 - The `arkworks` ecosystem [@arkworks] is probably the closest library, providing many of the
   necessary building blocks involved in Data Availability Sampling: prime fields, possibly paired
[…]
