### 2026-04-23 ad4a0ab
minor changes in paper.bib
