+++ paper.bib
@@ -17,11 +17,12 @@
     author  = {Boneh, Dan and Drake, Justin and Fisch, Ben and Gabizon, Ariel},
     journal = {Cryptology ePrint Archive},
     year    = {2020},
+    url     = {https://eprint.iacr.org/2020/081}
 }
 
 % 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
 @software{arkworks,
-  author = {arkworks contributors},
+  author = {{arkworks contributors}},
   title  = {\texttt{arkworks} zkSNARK ecosystem},
   url    = {https://arkworks.rs},
   year   = {2022},
@@ -47,7 +48,7 @@
 
 % 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
 @article{gabizon2019plonk,
-  title   = {Plonk: Permutations over lagrange-bases for oecumenical noninteractive arguments of knowledge},
+  title   = {Plonk: Permutations over {L}agrange-bases for oecumenical noninteractive arguments of knowledge},
   author  = {Gabizon, Ariel and Williamson, Zachary J and Ciobotaru, Oana},
   journal = {Cryptology ePrint Archive},
   year    = {2019},
@@ -89,7 +90,7 @@
 % 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
 @misc{rust-rse,
   author    = {{rust-rse contributors}},
-  title     = {Rust implementation of Reed-Solomon erasure coding},
+  title     = {Rust implementation of {R}eed-{S}olomon erasure coding},
   year      = {2021},
   publisher = {GitHub},
   journal   = {GitHub repository},
@@ -110,7 +111,7 @@
 
 % 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
 @article{stevan2024performance,
-    author  = {Stevan, A. and Lavaur, T. and Lacan, J. and Detchart, J. and Pérennou, T},
+    author  = {Stevan, Antoine and Lavaur, Thomas and Lacan, J{\'e}r{\^o}me and Detchart, Jonathan and P{\'e}rennou, Tanguy},
     title   = {Performance Evaluation of Polynomial Commitments for Erasure Code Based Information Dispersal},
     journal = {10th International Conference on Information Systems Security and Privacy},
     year    = {2024},
@@ -119,7 +120,7 @@
 
 @misc{rsmt2d,
   author    = {{rsmt2d contributors}},
-  title     = {Go implementation of two dimensional Reed-Solomon Merkle tree data availability scheme},
+  title     = {Go implementation of two dimensional {R}eed-{S}olomon {M}erkle tree data availability scheme},
   year      = {2026},
   publisher = {GitHub},
   journal   = {GitHub repository},
@@ -134,6 +135,7 @@
       archivePrefix={arXiv},
       primaryClass={cs.CR},
       url={https://arxiv.org/abs/1809.09044},
+      doi={10.48550/arXiv.1809.09044}
 }
 
 @draft{stevan2025komodo,
