### 2026-04-24 #9791 @a-stevan
i had a look at your PR, merged it into the `paper-joss` branch in [`gitlab.isae-supaero.fr:dragoon/komodo!210`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210) on GitLab (the mirror goes only one way) and just added [`6a3ded3d`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210/diffs?commit_id=6a3ded3d450c4da74aaca4b38598bcae075e6613) to fix what i thought were typos, let me know if they were not !

> - i was able to find _Efficient polynomial commitment schemes for multiple points and polynomials_ (`boneh2020efficient`) at https://eprint.iacr.org/2020/081
> - i was able to find _Fraud and Data Availability Proofs: Maximising Light Client Security and Scaling Blockchains with Dishonest Majorities_ (`albassam2019frauddataavailabilityproofs`) with DOI `10.48550/arXiv.1809.09044`
> - i'm fine with the other changes in `paper.bib`
> - i'm fine with the changes in `paper.md` => i just fixed what appeared to me as typos in [`6a3ded3d`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210/diffs?commit_id=6a3ded3d450c4da74aaca4b38598bcae075e6613)

at one point, it has been asked to copy the introduction of the paper to the README.
i removed it from the README some time ago to only sync it once when `paper.md` stops changing.
