### 2026-04-24 #9791 @danielskatz
@a-stevan

> do you still want us to copy the introduction from `paper.md` to `README.md` ?

This is up to you. Originally, I didn't think the README had enough general information about the project, which it now does, so if what's there is correct, it's fine to leave.
