### 2026-04-24 #9791 @danielskatz
@a-stevan - also I see that `gabizon2019plonk` should have
`url = {https://eprint.iacr.org/2019/953}`
added
