### 2026-04-24 #9791 @a-stevan
@danielskatz

got it !

i reverted the incorrect typos and added the URL to `paper.bib` in [`d352231`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210/diffs?commit_id=d35223183624500f06205d352d6e9aa1440347e6).

##### out of curiosity, i'm wondering why the `@kate2010constant` on line 56 is _direct_ but the `@gabizon2019plonk` on line 57 is _indirect_ ?

as for the [README](https://gitlab.isae-supaero.fr/dragoon/komodo/-/blob/paper-joss/README.md), i think the information is correct:
- it was correct before
- it has been improved thanks to @cassiersg i believe in
  - [`fa3bbabcb`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/fa3bbabcb3490425cf96f1d2780e7adc91384d81)
  - [`867ebaf62`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/867ebaf62a1bdbf4d3755e2f4d544fe05ad2291e)
  - [`4bc333af0`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/4bc333af0319390c90f7766e31890b9b27c19f8a)

**EDIT**: reading the README again, i thought adding the simplest example for the `./make.rs` command after the installation would be benefitial ([`16e8fe1`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/16e8fe1)).
