### 2026-04-24 #9791 @danielskatz
> out of curiosity, i'm wondering why the @kate2010constant on line 56 is direct but the @gabizon2019plonk on line 57 is indirect ?

@kate2010constant is the object of the wording on 56, while @gabizon2019plonk is not the object but a parenthetical reference following the object: `**PlonK**`

see https://joss.readthedocs.io/en/latest/paper.html#citations for more on this, and https://pandoc.org/MANUAL.html#extension-citations for even more
