### 2026-04-24 #9791 @a-stevan
> @kate2010constant is the object of the wording on 56, while @gabizon2019plonk is not the object but a parenthetical reference following the object: `**PlonK**`
>
> see https://joss.readthedocs.io/en/latest/paper.html#citations for more on this, and https://pandoc.org/MANUAL.html#extension-citations for even more
>
aaaaaah now i see, that was the page i was looking for ! thanks 🙏

its' the `i talk about @foo` (direct) vs `here i talk about foo [@foo]` (indirect)
