### 2026-04-24 #9791 @a-stevan
>  - Make a tagged release of your software, and list the version tag of the archived version here.
>
this should be straightforward.
my coworkers are currently in holydays, maybe i'll wait a bit to merge [!210](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210) into the main branch, if that's ok with you ?

i'll go ahead and mark [!210](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210) as ready for now.

> - Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> - Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
it's the first time i hear about Zenodo or similar, i'll see how it goes :)

> - Please list the DOI of the archived version here.
>
as soon as it's archived !

> I can then move forward with accepting the submission.
>
amazing, thanks a lot for the guidance and the patience 🙏
