### 2026-04-24 #9791 @danielskatz
I'm happy to wait a little bit.

Re the repositories, here's a list - https://www.re3data.org  Zenodo is the most commonly used one for JOSS software, but you can pick whichever you want.  (See https://help.zenodo.org/docs/github/ for some details on why and how - and https://github.com/zenodo/zenodo-rdm/issues/1203 seems to indicate this doesn't work as simply for GitLab though there is interest in making it so)
