### 2026-04-24 #9791 @a-stevan
i bumped the library to 1.2.0 in [`b1bf0cdc8`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210/diffs?commit_id=b1bf0cdc80dac2bf8699b6905af94ba8d121cfbd) because of the changes to the lib itself, i'll make a proper [release](https://gitlab.isae-supaero.fr/dragoon/komodo/-/releases) once [!210](https://gitlab.isae-supaero.fr/dragoon/komodo/-/merge_requests/210) lands.
