### 2026-04-29 #9791 @a-stevan
so we are still unclear about potential guidelines at my university (ISAE-SUPAERO, Toulouse, France) regarding publishing code, i.e. if they prefer Zenodo over HAL or something else...

i'll get back to you as soon as possible at the start of next week, the next two days are off here.
