### 2026-04-29 #9791 @danielskatz
We don't particularly care where the software is archived, as long as it is an archival repository that provides a DOI.
