### 2026-04-29 #9791 @a-stevan
> We don't particularly care where the software is archived, as long as it is an archival repository that provides a DOI.
>
perfect !
we have asked for an approved repository which provides a DOI, we are awaiting for their response.
