### 2026-05-04 29794d4
write a paper for the JOSS journal (dragoon/komodo!210)
This MR adds the JOSS paper of Komodo.

---

##### :green_circle: pre-review: [`github.com:openjournals/joss-reviews#9577`](https://github.com/openjournals/joss-reviews/issues/9577)
##### :green_circle: review: [`github.com:openjournals/joss-reviews#9791`](https://github.com/openjournals/joss-reviews/issues/9791)
##### :green_circle: tickets: [JOSS label](https://github.com/dragoon-rs/komodo/issues?q=label%3Ajoss) ([still opened](https://github.com/dragoon-rs/komodo/issues?q=label%3Ajoss%20is%3Aopen))

---

> [!note]
> template from https
[…]
