+++ paper.bib
@@ -1,0 +1,147 @@
+% 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
+@article{ambrona2023aplonk,
+    title        = {aPlonK: Aggregated PlonK from Multi-polynomial Commitment Schemes},
+    abstract     = {PlonK is a prominent universal and updatable zk-SNARK for general circuit satis ability. We present aPlonK, a variant of PlonK that reduces the proof size and veri cation time when multiple statements are proven in a batch. Both the aggregated proof size and the veri cation complexity of aPlonK are logarithmic in the number of aggregated statements. Our main building block, inspired by the techniques developed in SnarkPack (Gailly, Maller, Nitulescu, FC 2022), is a multi-polynomial commitment scheme, a new primitive that generalizes polynomial commitment schemes. Our techniques also include a mechanism for involving committed data into PlonK statements very e ciently, which can be of independent interest.},
+    language     = {en},
+    author       = {Ambrona, Miguel and Beunardeau, Marc and Schmitt, Anne-Laure and Toledo, Raphaël R},
+    booktitle    = {International Workshop on Security},
+    pages        = {195--213},
+    year         = {2023},
+    organization = {Springer},
+    doi          = {10.1007/978-3-031-41326-1_11},
+}
+
+% 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
+@article{boneh2020efficient,
+    title   = {Efficient polynomial commitment schemes for multiple points and polynomials},
+    author  = {Boneh, Dan and Drake, Justin and Fisch, Ben and Gabizon, Ariel},
+    journal = {Cryptology ePrint Archive},
+    year    = {2020},
+    url     = {https://eprint.iacr.org/2020/081}
+}
+
+% 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
+@software{arkworks,
+  author = {{arkworks contributors}},
+  title  = {\texttt{arkworks} zkSNARK ecosystem},
+  url    = {https://arkworks.rs},
+  year   = {2022},
+}
+
+% 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
+@article{ethereum2024danksharding,
+  title = {What is Proto-Danksharding?},
+  url = {https://ethereum.org/roadmap/danksharding/#what-is-protodanksharding},
+  year = {2024},
+}
+
+% 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
+@inproceedings{fiat1986prove,
+    title        = {How to prove yourself: Practical solutions to identification and signature problems},
+    author       = {Fiat, Amos and Shamir, Adi},
+    booktitle    = {Conference on the theory and application of cryptographic techniques},
+    pages        = {186--194},
+    year         = {1986},
+    organization = {Springer},
+    doi          = {10.1007/3-540-47721-7_12},
+}
+
+% 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
+@article{gabizon2019plonk,
+  title   = {Plonk: Permutations over {L}agrange-bases for oecumenical noninteractive arguments of knowledge},
+  author  = {Gabizon, Ariel and Williamson, Zachary J and Ciobotaru, Oana},
+  journal = {Cryptology ePrint Archive},
+  year    = {2019},
+  url     = {https://eprint.iacr.org/2019/953},
+}
+
+% 6ff062973af6e9965071d81ab7fb3b6128c142eee89f3f79c847a0ba50e0030c
+@inproceedings{kate2010constant,
[…]
