+++ paper.md
@@ -1,0 +1,148 @@
+---
+title: 'Komodo: Cryptographically-proven Erasure Coding'
+tags:
+  - Rust
+  - cryptography
+  - erasure codes
+  - distributed systems
+  - data availability sampling
+authors:
+  - name: Antoine Stevan
+    orcid: 0009-0003-5684-5862
+    equal-contrib: true
+    affiliation: 1
+  - name: Jonathan Detchart
+    orcid: 0000-0002-4237-5981
+    equal-contrib: true
+    affiliation: 1
+  - name: Tanguy Pérennou
+    orcid: 0009-0002-2542-0004
+    equal-contrib: false
+    affiliation: 1
+  - name: Jérôme Lacan
+    orcid: 0000-0002-3121-4824
+    equal-contrib: false
+    affiliation: 1
+affiliations:
+ - name: ISAE-SUPAERO, France
+   index: 1
+date: 24 November 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+We present **Komodo**, a library that allows one to (**1**) encode data with forward erasure-code (FEC) techniques such as Reed-Solomon encoding, (**2**) prove the resulting _shards_ with cryptographic protocols, (**3**) verify their validity on the other end of any distributed network, and (**4**) decode the original data from a subset of said _shards_ [@stevan2024performance; @stevan2023assessing].
+The library is implemented in the _Rust_ programming language and available under the MIT license on the ISAE-SUPAERO GitLab instance[^1] with a mirror on GitHub[^2].
+**Komodo** should be of interest for people willing to explore the field of cryptographically-proven _shards_ of data in distributed systems or data availability sampling settings.
+
+A _shard_ of data is said to be _valid_ if it has been generated by a correct FEC encoding of the data.
+
+**Komodo** implements a protocol API to achieve the following on any input data in a distributed network:
+
+- `encode`: encodes data into _shards_ with a Reed-Solomon $(k,n)$-code.
+  This adds redundancy to the data, making the network more resilient to failure, fragmentation, partitioning, loss or corruption.
+- `commit`: commit the input data into a _commitment_ common to all _shards_.
+- `prove`: generate one proof for all $n$ _shards_ with one of three available cryptographic protocols (see below for more information).
+  This extra information guarantees a given _shard_ is valid, except with negligible probability.
+- `verify`: verifies a _shard_ individually for its validity using the data _commitment_ and the _shard_ proof.
+  This allows to discriminate invalid or corrupted _shards_ without any decoding attempt.
+  Without this shard-level verification step, it is impossible to know if a _shard_ is valid until the decoding step.
+  Then, when decoding fails, it is not possible to know which _shards_ were invalid, leading to a _try-and-error_ process that is not scalable.
+- `decode`: decodes the original data using any subset of $k$ independent and valid _shards_.
+
+**Komodo** provides the three following cryptographic protocols:
+
+- **KZG+** (short for the first letters of the authors' names): based on @kate2010constant and its multi-polynomial extension [@boneh2020efficient]
+- **aPlonK** (short for _**A**ggregated **P**ermutations over **L**agrange-bases for **O**ecumenical **N**oninteractive arguments of **K**nowledge_): based on **PlonK** [@gabizon2019plonk] and **aPlonK** [@ambrona2023aplonk]
+- **Semi-AVID** (short for _Semi **A**synchronous **V**erifiable **I**nformation **D**ispersal_): based on **Semi-AVID-PR** [@nazirkhanova2022information]
+
[…]
