### 2026-05-04 #9791 @a-stevan
@danielskatz 👋

merge request 210 has just landed, i tagged the code: [1.2.0](https://gitlab.isae-supaero.fr/dragoon/komodo/-/tags/1.2.0)

we have made an archive on hal.science just now with a tarball of [29794d4](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/29794d464799949b0dd80f7c6b397942dfce0f2f) and are _waiting for moderation_.
