### 2026-05-05 5f190f2
remove useless code from kzg::prove (dragoon/komodo!244)
should close #29
