### 2026-05-06 #9791 @a-stevan
well https://github.com/dragoon-rs/komodo remains a mirror of all the protected branches from https://gitlab.isae-supaero.fr/dragoon/komodo
- `main`
- `v1.0.2`
- `v1.1.0`

`paper-joss` was a protected MR branch for !210 which was merged and thus the branch has been deleted.

now, the `main` branch will continue growing but the 1.2.0 tag will remain on the MR commit [`29794d4`](https://gitlab.isae-supaero.fr/dragoon/komodo/-/commit/29794d464799949b0dd80f7c6b397942dfce0f2f) .
