### 2026-05-11 29bf7b6
post JOSS 10.21105/joss.09791 (dragoon/komodo!245)
> [!important]
> !210 has landed and [`github.com:openjournals/joss-reviews#9791`](https://github.com/openjournals/joss-reviews/issues/9791) was just accepted :tada: 
>
> ==> Komodo now has a DOI: [**10.21105/joss.09791**](https://joss.theoj.org/papers/10.21105/joss.09791)

This MR implements two suggestions from [@editorialbot]
- add `CITATION.cff` for people to cite the JOSS paper: [`github.com:openjournals/joss-reviews#9791` `4388582292`](https://github.com/openjournals/joss-reviews/issues/9791#issuecomment-4388582292)
- add a DOI badge to 
[…]
